Refactor to make sharing JS easier
