+++ src/js/client/payment.js
