+++ src/js/client/polyfills.js
