+++ src/resources/workshops.yml
