+++ src/styles/payment.styl
