+++ src/views/terms.html
